fix: Refactored the project structure
