+++ java/src/main/java/com/codurance/training/tasks/controller/AddProject.java
@@ -1,4 +1,7 @@
-package com.codurance.training.tasks;
+package com.codurance.training.tasks.controller;
+
+import com.codurance.training.tasks.entity.Project;
+import com.codurance.training.tasks.service.AddService;
 
 import java.io.PrintWriter;
 import java.util.Map;
