+++ java/src/main/java/com/codurance/training/tasks/entity/Project.java
@@ -1,4 +1,4 @@
-package com.codurance.training.tasks;
+package com.codurance.training.tasks.entity;
 
 import java.util.ArrayList;
 import java.util.List;
