+++ java/src/main/java/com/codurance/training/tasks/entity/Task.java
@@ -1,4 +1,4 @@
-package com.codurance.training.tasks;
+package com.codurance.training.tasks.entity;
 
 import java.util.Date;
