+++ java/src/main/java/com/codurance/training/tasks/service/StatusService.java
@@ -1,4 +1,6 @@
-package com.codurance.training.tasks;
+package com.codurance.training.tasks.service;
+
+import com.codurance.training.tasks.entity.Project;
 
 import java.io.PrintWriter;
 import java.util.Map;
